Refactor to use DTO instead inner class
